refactor: lib ディレクトリを src/core へ移動
