+++ src/core/bgPhysics.js
