+++ src/core/bootOverlay.js
