+++ src/core/controls.js
@@ -1,4 +1,4 @@
-// lib/controls.js
+// src/core/controls.js
 // OrbitControls を薄くラップし、ズームやパンを禁止する
 import { OrbitControls } from "https://unpkg.com/three@0.160.0/examples/jsm/controls/OrbitControls.js";
