+++ src/core/data/works.js
@@ -1,4 +1,4 @@
-// lib/data/works.js
+// src/core/data/works.js
 // WORKS データを定義
 // 現状は日本語のみだが、将来的に多言語や JSON 取得へ拡張できる構造
