+++ src/entry/avatar.js
@@ -4,14 +4,14 @@
 // - ブートオーバーレイとリサイズ処理を管理
 import * as THREE from "https://unpkg.com/three@0.160.0/build/three.module.js";
 
-import { CONFIG } from "../../lib/config.js";
-import { createControls } from "../../lib/controls.js";
-import { createPostPipeline } from "../../lib/postprocess.js";
+import { CONFIG } from "../core/config.js";
+import { createControls } from "../core/controls.js";
+import { createPostPipeline } from "../core/postprocess.js";
 import { createRenderer,
-  fitToCanvas } from "../../lib/renderer.js";
-import { createSceneGraph } from "../../lib/scene.js";
-import { applyPS1Jitter, runFixedStepLoop } from "../../lib/utils.js";
-import { initBootOverlay } from "../../lib/bootOverlay.js";
+  fitToCanvas } from "../core/renderer.js";
+import { createSceneGraph } from "../core/scene.js";
+import { applyPS1Jitter, runFixedStepLoop } from "../core/utils.js";
+import { initBootOverlay } from "../core/bootOverlay.js";
 
 const canvas = document.getElementById("avatar-canvas");
 const renderer = createRenderer(THREE, canvas, CONFIG);
