+++ src/entry/background.js
@@ -4,9 +4,9 @@
 // - 壁との反射と球同士の弾性衝突を実装
 
 import * as THREE from "https://unpkg.com/three@0.160.0/build/three.module.js";
-import { createRenderer, fitToCanvas } from "../../lib/renderer.js";
-import { runFixedStepLoop } from "../../lib/utils.js";
-import { BgPhysics } from "../../lib/bgPhysics.js";
+import { createRenderer, fitToCanvas } from "../core/renderer.js";
+import { runFixedStepLoop } from "../core/utils.js";
+import { BgPhysics } from "../core/bgPhysics.js";
 
 const canvas = document.getElementById("bg-canvas");
 if (!canvas)
