+++ src/entry/works-gallery.js
@@ -1,7 +1,7 @@
 // works-gallery.js
 // 制作物ギャラリーのタブ切替と表示処理
 // WORKS データを別モジュールから読み込む
-import { WORKS } from '../../lib/data/works.js';
+import { WORKS } from '../core/data/works.js';
 
 const tabs = document.querySelectorAll('.works-tabs button');
 const image = document.querySelector('.works-image img');
